Split list in two
I split the linked list at the midpoint.
In an list with odd number of nodes, the first
list from the first half will always be shorter
than the second list obtained from the second half.
In a list with even number of nodes, both list will
always be equal.